+++ retrocache/src/test/java/com/andiag/retrocache/cache/ToStringConverterFactory.java
@@ -18,11 +18,13 @@ public class ToStringConverterFactory extends Converter.Factory {
     private static final MediaType MEDIA_TYPE = MediaType.parse("text/plain");
 
     @Override
-    public Converter<ResponseBody, ?> responseBodyConverter(Type type, Annotation[] annotations,
+    public Converter<ResponseBody, ?> responseBodyConverter(Type type,
+                                                            Annotation[] annotations,
                                                             Retrofit retrofit) {
         if (String.class.equals(type)) {
             return new Converter<ResponseBody, String>() {
-                @Override public String convert(ResponseBody value) throws IOException {
+                @Override
+                public String convert(ResponseBody value) throws IOException {
                     return value.string();
                 }
             };
@@ -30,11 +32,17 @@ public class ToStringConverterFactory extends Converter.Factory {
         return null;
     }
 
-    @Override public Converter<?, RequestBody> requestBodyConverter(Type type,
-                                                                    Annotation[] parameterAnnotations, Annotation[] methodAnnotations, Retrofit retrofit) {
+    @Override
+    public Converter<?, RequestBody> requestBodyConverter(
+            Type type,
+            Annotation[] parameterAnnotations,
+            Annotation[] methodAnnotations,
+            Retrofit retrofit) {
+
         if (String.class.equals(type)) {
             return new Converter<String, RequestBody>() {
-                @Override public RequestBody convert(String value) throws IOException {
+                @Override
+                public RequestBody convert(String value) throws IOException {
                     return RequestBody.create(MEDIA_TYPE, value);
                 }
             };